+++ src/Database/001_CreateDatabase.sql
@@ -1,85 +1,3 @@
-USE [master]
-GO
-/****** Object:  Database [PianoLesson]    Script Date: 07-Nov-25 8:47:03 AM ******/
-CREATE DATABASE [PianoLesson]
- CONTAINMENT = NONE
- ON  PRIMARY 
-( NAME = N'PianoLesson', FILENAME = N'D:\SQL Server DB\MSSQL16.SQLEXPRESS\MSSQL\DATA\PianoLesson.mdf' , SIZE = 8192KB , MAXSIZE = UNLIMITED, FILEGROWTH = 65536KB )
- LOG ON 
-( NAME = N'PianoLesson_log', FILENAME = N'D:\SQL Server DB\MSSQL16.SQLEXPRESS\MSSQL\DATA\PianoLesson_log.ldf' , SIZE = 8192KB , MAXSIZE = 2048GB , FILEGROWTH = 65536KB )
- WITH CATALOG_COLLATION = DATABASE_DEFAULT, LEDGER = OFF
-GO
-ALTER DATABASE [PianoLesson] SET COMPATIBILITY_LEVEL = 160
-GO
-IF (1 = FULLTEXTSERVICEPROPERTY('IsFullTextInstalled'))
-begin
-EXEC [PianoLesson].[dbo].[sp_fulltext_database] @action = 'enable'
-end
-GO
-ALTER DATABASE [PianoLesson] SET ANSI_NULL_DEFAULT OFF 
-GO
-ALTER DATABASE [PianoLesson] SET ANSI_NULLS OFF 
-GO
-ALTER DATABASE [PianoLesson] SET ANSI_PADDING OFF 
-GO
-ALTER DATABASE [PianoLesson] SET ANSI_WARNINGS OFF 
-GO
-ALTER DATABASE [PianoLesson] SET ARITHABORT OFF 
-GO
-ALTER DATABASE [PianoLesson] SET AUTO_CLOSE OFF 
-GO
-ALTER DATABASE [PianoLesson] SET AUTO_SHRINK OFF 
-GO
-ALTER DATABASE [PianoLesson] SET AUTO_UPDATE_STATISTICS ON 
-GO
-ALTER DATABASE [PianoLesson] SET CURSOR_CLOSE_ON_COMMIT OFF 
-GO
-ALTER DATABASE [PianoLesson] SET CURSOR_DEFAULT  GLOBAL 
-GO
-ALTER DATABASE [PianoLesson] SET CONCAT_NULL_YIELDS_NULL OFF 
-GO
-ALTER DATABASE [PianoLesson] SET NUMERIC_ROUNDABORT OFF 
-GO
-ALTER DATABASE [PianoLesson] SET QUOTED_IDENTIFIER OFF 
-GO
-ALTER DATABASE [PianoLesson] SET RECURSIVE_TRIGGERS OFF 
-GO
-ALTER DATABASE [PianoLesson] SET  DISABLE_BROKER 
-GO
-ALTER DATABASE [PianoLesson] SET AUTO_UPDATE_STATISTICS_ASYNC OFF 
-GO
-ALTER DATABASE [PianoLesson] SET DATE_CORRELATION_OPTIMIZATION OFF 
-GO
-ALTER DATABASE [PianoLesson] SET TRUSTWORTHY OFF 
-GO
-ALTER DATABASE [PianoLesson] SET ALLOW_SNAPSHOT_ISOLATION OFF 
-GO
-ALTER DATABASE [PianoLesson] SET PARAMETERIZATION SIMPLE 
-GO
-ALTER DATABASE [PianoLesson] SET READ_COMMITTED_SNAPSHOT OFF 
-GO
-ALTER DATABASE [PianoLesson] SET HONOR_BROKER_PRIORITY OFF 
-GO
-ALTER DATABASE [PianoLesson] SET RECOVERY SIMPLE 
-GO
-ALTER DATABASE [PianoLesson] SET  MULTI_USER 
-GO
-ALTER DATABASE [PianoLesson] SET PAGE_VERIFY CHECKSUM  
-GO
-ALTER DATABASE [PianoLesson] SET DB_CHAINING OFF 
-GO
-ALTER DATABASE [PianoLesson] SET FILESTREAM( NON_TRANSACTED_ACCESS = OFF ) 
-GO
-ALTER DATABASE [PianoLesson] SET TARGET_RECOVERY_TIME = 60 SECONDS 
-GO
-ALTER DATABASE [PianoLesson] SET DELAYED_DURABILITY = DISABLED 
-GO
-ALTER DATABASE [PianoLesson] SET ACCELERATED_DATABASE_RECOVERY = OFF  
-GO
-ALTER DATABASE [PianoLesson] SET QUERY_STORE = ON
-GO
-ALTER DATABASE [PianoLesson] SET QUERY_STORE (OPERATION_MODE = READ_WRITE, CLEANUP_POLICY = (STALE_QUERY_THRESHOLD_DAYS = 30), DATA_FLUSH_INTERVAL_SECONDS = 900, INTERVAL_LENGTH_MINUTES = 60, MAX_STORAGE_SIZE_MB = 1000, QUERY_CAPTURE_MODE = AUTO, SIZE_BASED_CLEANUP_MODE = AUTO, MAX_PLANS_PER_QUERY = 200, WAIT_STATS_CAPTURE_MODE = ON)
-GO
 USE [PianoLesson]
 GO
 /****** Object:  Table [dbo].[Lesson]    Script Date: 07-Nov-25 8:47:03 AM ******/
