+++ src/MusicPiano/MusicPianoWebAPI/Controllers/WeatherForecastController.cs
@@ -1,0 +1,47 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace MusicPianoWebAPI.Controllers
+{
+    [ApiController]
+    [Route("api/[controller]/[action]")]
+    public class WeatherForecastController : ControllerBase
+    {
+        private static readonly string[] Summaries = new[]
+        {
+            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+        };
+
+        private readonly ILogger<WeatherForecastController> _logger;
+
+        public WeatherForecastController(ILogger<WeatherForecastController> logger)
+        {
+            _logger = logger;
+        }
+
+        [HttpGet()]
+        public IEnumerable<WeatherForecast> GetData()
+        {
+            return Enumerable.Range(1, 5).Select(index => new WeatherForecast
+            {
+                Date = DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
+                TemperatureC = Random.Shared.Next(-20, 55),
+                Summary = Summaries[Random.Shared.Next(Summaries.Length)]
+            })
+            .ToArray();
+        }
+
+        [HttpGet()]
+        public string HelloWorld()
+        {
+            var day = DateTime.Now.Minute;
+            if (day % 2 == 0)
+            {
+                return "Hello World";
+            }
+            else
+            {
+                throw new Exception("Not a current date");
+            }
+        }
+    }
+}
