+++ src/MusicPiano/MusicPianoWebAPI/MusicPianoWebAPI.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.15" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MusicPianoBusinessLogic\MusicPianoBusinessLogic.csproj" />
+    <ProjectReference Include="..\MusicPianoData\MusicPianoData.csproj" />
+  </ItemGroup>
+
+</Project>
