+++ src/MusicPiano/MusicPianoWebAPI/WeatherForecast.cs
@@ -1,0 +1,13 @@
+namespace MusicPianoWebAPI
+{
+    public class WeatherForecast
+    {
+        public DateOnly Date { get; set; }
+
+        public int TemperatureC { get; set; }
+
+        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+
+        public string? Summary { get; set; }
+    }
+}
