+++ src/MusicPiano/MusicPianoWebAPI/appsettings.json
@@ -1,0 +1,12 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "PianoDatabase": "Data Source=.;Initial Catalog=PianoLesson;TrustServerCertificate=True;User ID=SA;Password=Faringeal2!;"
+  }
+}
